finance/azure-search/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#azure-functions
azure.functions~=1.0a
langchain==0.1.0
openai==1.7.0
